opengraph implementation and remove impport all React to specific useEffect + useState
