+++ src/components/ActiveArticlesList.tsx
@@ -51,8 +51,12 @@ const ActiveArticlesList: React.FC<ActiveArticlesProps> = ({
                                             <Image
                                                 src={article.banner}
                                                 alt={`Banner for ${article.title}`}
-                                                layout='fill'
-                                                objectFit='cover'
+                                                fill
+                                                priority={true}
+                                                style={{
+                                                    objectFit: 'cover',
+                                                }}
+                                                sizes='(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw'
                                                 className='rounded-md'
                                             />
                                         </div>
